backend/requirements.txt
fastapi
uvicorn
sqlalchemy
mysqlclient
pydantic
python-dotenv
passlib[bcrypt]
aiohttp
python-jose
